+++ src/main/java/sorts/qsort/Qsort.java
@@ -31,7 +31,7 @@ public class Qsort extends AbstractSorter implements Time {
      * Return measured time
      */
     @Override
-    public long getTime() {
+    public double getTime() {
         //System.out.print("Time is " + time);
         return time;
     }
